refactor(gitinfo): split god file into per-concern components
Split internal/panels/gitinfo/gitinfo.go (~5000 lines) into focused
files by domain: branches, worktrees, remotes, stash, tags, reflog,
and GitHub integration. No behavior changes.

Resolves #64
